// Client/src/ServerLibs/SocketType.ts
enum SocketType {
    Authenticator = "Authenticator",
    AnonymousGameManager = "AnonymousGameManager",
    AnonymousMatchmaker = "AnonymousMatchmaker",
    AnonymousRoomManager = "AnonymousRoomManager",
    GameManager = "GameManager",
    Matchmaker = "Matchmaker",
    RoomManager = "RoomManager",
}

export { SocketType }